requirements.txt
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-openai==0.2.10
langchain-text-splitters==0.3.0
langsmith==0.1.137
PyMuPDF==1.24.14
pypdf==5.0.1
PyPDF2==3.0.1
huggingface==0.0.1
chromadb==0.5.20
python-dotenv==1.0.1